examples/taco_tweets/requirements.txt
clarifai
Pillow
requests
requests_oauthlib